requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart
# Vector Database
chromadb>=0.4.0
# LLM & DSPy
dspy-ai>=2.1.0
ollama>=0.1.6
# Search & Math
rank_bm25
numpy
# Machine Learning (Re-ranking)
sentence-transformers
torch
# Utils
python-magic-bin; sys_platform == 'win32'
python-magic; sys_platform != 'win32'
structlog
redis
celery
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
prometheus-client
tenacity 
pypdf